// react/src/core/publizon/model/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Publizon Library API
 * Pubhub exists in two separate environments, each with their own server, code and database. Please use the web service located at library-api.qa.pubhub.dk when developing and testing.
 * Orders/loans created in test environment will not be invoiced. Please request a new license key for the production environment when you're ready to go live.
 * OpenAPI spec version: 1
 */
import type { LibraryProductCategory } from "./libraryProductCategory";
import type { LibraryProductContributor } from "./libraryProductContributor";
import type { LibraryProductExternalId } from "./libraryProductExternalId";

export interface Product {
  createdUtc?: string;
  updatedUtc?: string;
  /** @nullable */
  title?: string | null;
  isActive?: boolean;
  /** @nullable */
  languageCode?: string | null;
  /** @nullable */
  coverUri?: string | null;
  /** @nullable */
  thumbnailUri?: string | null;
  /** @nullable */
  sampleUri?: string | null;
  productType?: number;
  externalProductId?: LibraryProductExternalId;
  /** @nullable */
  internalProductId?: string | null;
  /** @nullable */
  contributors?: LibraryProductContributor[] | null;
  /** @nullable */
  format?: string | null;
  fileSizeInBytes?: number;
  /** @nullable */
  durationInSeconds?: number | null;
  /** @nullable */
  publisher?: string | null;
  publicationDate?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  productCategories?: LibraryProductCategory[] | null;
  costFree?: boolean;
}