// react/src/core/publizon/model/loan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Publizon Library API
 * Pubhub exists in two separate environments, each with their own server, code and database. Please use the web service located at library-api.qa.pubhub.dk when developing and testing.
 * Orders/loans created in test environment will not be invoiced. Please request a new license key for the production environment when you're ready to go live.
 * OpenAPI spec version: 1
 */
import type { FileExtensionType } from "./fileExtensionType";
import type { LibraryBook } from "./libraryBook";

export interface Loan {
  orderId?: string;
  /** @nullable */
  orderNumber?: string | null;
  orderDateUtc?: string;
  loanExpireDateUtc?: string;
  isSubscriptionLoan?: boolean;
  libraryBook?: LibraryBook;
  fileExtensionType?: FileExtensionType;
}